requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
openai==1.65.2
anthropic==0.49.0
google-generativeai==0.8.4
requests==2.32.3